+++ source/core/aggregator-manager/WorkerService.h
@@ -23,7 +23,7 @@ namespace sf1r
 {
 using izenelib::ir::idmanager::IDManager;
 
-typedef std::vector<std::pair<uint32_t, float> > SimilarDocIdListType;
+typedef std::vector<std::pair<uint64_t, float> > SimilarDocIdListType;
 
 class IndexBundleConfiguration;
 class IndexSearchService;
@@ -62,7 +62,7 @@ public:
     bool getInternalDocumentId(const izenelib::util::UString& actionItem, uint64_t& resultItem);
 
     /// mining
-    bool getSimilarDocIdList(uint32_t& documentId, uint32_t& maxNum, SimilarDocIdListType& result);
+    bool getSimilarDocIdList(uint64_t& documentId, uint32_t& maxNum, SimilarDocIdListType& result);
 
     bool clickGroupLabel(const ClickGroupLabelActionItem& actionItem, bool& ret);
